requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
pydantic==2.10.6
pymongo==4.11.1
python-dotenv==1.0.1
python-multipart==0.0.20
razorpay==1.4.2
requests==2.32.3
passlib==1.7.4
PyJWT==2.10.1
email_validator==2.2.0
setuptools==59.6.0